fix: architectural changes & microsoft fixes
